pyproject.toml
[tool.poetry]
name = "modern-dev"
version = "0.1.0"
description = ""
authors = ["Guillermo Ramirez Garcia <memo@hpe.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.23.0"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
uvloop = "^0.17.0"
pydantic = "^1.10.4"
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
opentelemetry-instrumentation-fastapi = "^0.36b0"
starlette-exporter = "^0.14.0"
opentelemetry-distro = "^0.36b0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
pyright = "^1.1.286"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 95
target-version = ["py311"]


[tool.pyright]
include = ["src"]
exclude = [
    "**/node_modules",
    "**/__pycache__"
]
# ignore = ["src/oldstuff"]
# stubPath = "src/stubs"
# venv = ".venv"

defineConstant = { DEBUG = true }

pythonVersion = "3.11"
pythonPlatform = "Linux"

executionEnvironments = []

reportMissingImports = true
# reportMissingTypeStubs = true
strictDictionaryInference = true
strictListInference = true
analyzeUnannotatedFunctions = true
strictParameterNoneValue = true
enableTypeIgnoreComments = true
reportMissingModuleSource = "warning"
reportUndefinedVariable = "error"
reportAssertAlwaysTrue = "error"
reportInvalidStringEscapeSequence = "error"
reportInvalidTypeVarUse = "error"
reportSelfClsParameterName = "error"
reportUnsupportedDunderAll = "error"
reportUnusedExpression = "error"
reportWildcardImportFromLibrary = "error"
reportGeneralTypeIssues = "error"
reportOptionalSubscript = "error"
reportOptionalMemberAccess = "error"
reportOptionalCall = "error"
reportOptionalIterable = "error"
reportOptionalContextManager = "error"
reportOptionalOperand = "error"
reportTypedDictNotRequiredAccess = "error"
reportPrivateImportUsage = "error"
reportUnboundVariable = "error"
reportUnusedCoroutine = "error"
reportConstantRedefinition = "error"
reportDuplicateImport = "error"
reportFunctionMemberAccess = "error"
reportImportCycles = "error"
reportIncompatibleMethodOverride = "error"
reportIncompatibleVariableOverride = "error"
# reportIncompleteStub = "error"
reportInconsistentConstructor = "error"
# reportInvalidStubStatement = "error"
reportMatchNotExhaustive = "error"
reportMissingParameterType = "error"
reportMissingTypeArgument = "error"
reportOverlappingOverload = "error"
reportPrivateUsage = "error"
reportShadowedImports = "none"
reportTypeCommentUsage = "error"
reportUnknownArgumentType = "error"
reportUnknownLambdaType = "error"
# reportUnknownMemberType = "error"
reportUnknownParameterType = "error"
reportUnknownVariableType = "error"
reportUnnecessaryCast = "error"
reportUnnecessaryComparison = "error"
reportUnnecessaryContains = "error"
reportUnnecessaryIsInstance = "error"
reportUnusedClass = "error"
reportUnusedImport = "error"
reportUnusedFunction = "error"
reportUnusedVariable = true
reportUntypedBaseClass = "error"
reportUntypedClassDecorator = "error"
reportUntypedFunctionDecorator = "error"
reportUntypedNamedTuple = "error"
reportCallInDefaultInitializer = "none"
reportImplicitStringConcatenation = "none"
reportMissingSuperCall = "none"
reportPropertyTypeMismatch = "none"
reportUninitializedInstanceVariable = "none"
reportUnnecessaryTypeIgnoreComment = "none"
reportUnusedCallResult = "none"
